+++ src/include/Utilities.h
@@ -9,12 +9,12 @@
 #include <complex.h>
 
 
-float **init_2DMatrix(const unsigned int nRows, const unsigned int nColumns);
-void free_2DMatrix(float **matrix, const unsigned int nRows);
-bool is_in_row(float value, float *array, const unsigned int size);
-bool is_in_matrix(float *value, float **matrix, const unsigned int size);
-void dot_product(float m1[][dimension], float m2[][dimension], float mp[][dimension]);
-void matrix_transpose(float m[][dimension], float mt[][dimension]);
-void eigen_values(float matrix[][dimension], float eigenvalues[]);
+double **init_2DMatrix(const unsigned int nRows, const unsigned int nColumns);
+void free_2DMatrix(double **matrix, const unsigned int nRows);
+bool is_in_row(double value, double *array, const unsigned int size);
+bool is_in_matrix(double *value, double **matrix, const unsigned int size);
+void dot_product(double m1[][dimension], double m2[][dimension], double mp[][dimension]);
+void matrix_transpose(double m[][dimension], double mt[][dimension]);
+void eigen_values(double matrix[][dimension], double eigenvalues[]);
 
 #endif
